requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
pydantic==1.10.7
python-dotenv==1.0.0
imaplib2==3.6
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
PyGithub==1.58.1
openai==0.27.4
pytest==7.3.1
pytest-mock==3.10.0
python-multipart==0.0.6
email-validator==2.0.0
jinja2==3.1.2
psycopg2-binary==2.9.6
